Remove graphic debug information
Casted something in random, though it does nothing

+++ lib/math/random.c
@@ -2,7 +2,7 @@
 #include "random.h"
 
 double uniform() {
-    return rand() / (RAND_MAX + 1.0);
+    return (float)rand() / (RAND_MAX + 1.0);
 }
 
 double rand_between(double min, double max) {
